Update README.md
TASK 1: NUMBER GAME:
Generate a random number within a specified range, such as 1 to 100.
Prompt the user to enter their guess for the generated number.
Compare the user's guess with the generated number and provide feedback on whether the guess is correct, too high, or too low.
Repeat steps 2 and 3 until the user guesses the correct number. You can incorporate additional details as follows:
Limit the number of attempts the user has to guess the number.
Add the option for multiple rounds, allowing the user to play again.
Display the user's score, which can be based on the number of attempts taken or rounds won.


TASK 2: STUDENT GRADE CALCULATOR:
Take marks obtained (out of 100) in each subject. Calculate Total Marks: Sum up the marks obtained in all subjects. Calculate Average Percentage: Divide the total marks by the total number of subjects to get the average percentage. Grade Calculation: Assign grades based on the average percentage achieved. Display Results: Show the total marks, average percentage, and the corresponding grade to the user

TASK 3: ATM INTERFACE:
Create a class to represent the ATM machine. 2. Design the user interface for the ATM, including options such as withdrawing, depositing, and checking the balance. 3. Implement methods for each option, such as withdraw(amount), deposit(amount), and checkBalance(). 4. Create a class to represent the user's bank account, which stores the account balance. 5. Connect the ATM class with the user's bank account class to access and modify the account balance. 6. Validate user input to ensure it is within acceptable limits (e.g., sufficient balance for withdrawals). 7. Display appropriate messages to the user based on their chosen options and the success or failure of their transactions.